Finalizing tests and validations

+++ Application/Services/Devices/DeviceService.cs
@@ -24,58 +24,92 @@ namespace OneGlobal.Application.Services.Devices
 
         public async Task<IEnumerable<Device>> GetDevicesByStateAsync(State state)
         {
-            throw new NotImplementedException();
+            var devices = await _deviceRepository.GetAllAsync();
+            return devices.Where(d => d.State == state);
         }
 
         public async Task<IEnumerable<Device>> GetDevicesByBrandAsync(string brand)
         {
-            throw new NotImplementedException();
+            if (string.IsNullOrWhiteSpace(brand))
+            {
+                throw new ArgumentNullException(nameof(brand));
+            }
+            var devices = await _deviceRepository.GetAllAsync();
+            return devices.Where(d => d.Brand.ToLower() == brand.ToLower());
         }
 
         public async Task<Device> GetDeviceByIdAsync(Guid id)
         {
+            if (id == Guid.Empty)
+            {
+                throw new ArgumentNullException(nameof(id));
+            }
             return await _deviceRepository.GetByIdAsync(id);
         }
 
         public async Task<Device> AddDeviceAsync(Device device)
         {
+            if (device == null)
+            {
+                throw new ArgumentNullException(nameof(device));
+            }
+            await ValidateDeviceForAdd(device);
             return await _deviceRepository.AddAsync(device);
         }
 
         public async Task<Device> UpdateDeviceAsync(Guid id, DevicePatch devicePatch)
         {
+            if (id == Guid.Empty)
+            {
+                throw new ArgumentNullException(nameof(id));
+            }
+            if (devicePatch == null)
+            {
+                throw new ArgumentNullException(nameof(devicePatch));
+            }
             await ValidateDeviceForUpdate(id, devicePatch);
             return await _deviceRepository.UpdateAsync(id, devicePatch);
         }
 
         public async Task<Device> UpdateDevicePartialAsync(Guid id, DevicePatch devicePatch)
         {
+            if (id == Guid.Empty)
+            {
+                throw new ArgumentNullException(nameof(id));
+            }
+            if (devicePatch == null)
+            {
+                throw new ArgumentNullException(nameof(devicePatch));
+            }
             await ValidateDeviceForUpdate(id, devicePatch);
             return await _deviceRepository.UpdatePartialAsync(id, devicePatch);
         }
 
         public async Task DeleteDeviceAsync(Guid id)
         {
+            if (id == Guid.Empty)
+            {
+                throw new ArgumentNullException(nameof(id));
+            }
             await ValidateDeviceForDelete(id);
             await _deviceRepository.DeleteAsync(id);
+        }
+
+        private async Task ValidateDeviceForAdd(Device device)
+        {
+            DeviceValidations.IsValidForAdd(device);
         }
 
         private async Task ValidateDeviceForUpdate(Guid id, DevicePatch patch)
         {
             var current = await _deviceRepository.GetByIdAsync(id);
-            if (!DeviceValidations.IsValidForUpdate(current, patch))
-            {
-                throw new InvalidStateForUpdateException(current.Id);
-            }
+            DeviceValidations.IsValidForUpdate(current, patch);
         }
 
         private async Task ValidateDeviceForDelete(Guid id)
         {
             var device = await _deviceRepository.GetByIdAsync(id);
-            if (!DeviceValidations.IsValidForDelete(device))
-            {
-                throw new InvalidStateForDeleteException(device.Id);
-            }
+            DeviceValidations.IsValidForDelete(device);
         }
     }
 }
